fix: do not provide fake values
The decision taken to provide wrong or fake values
is not acceptable, for example `0000000000000000000` is not a valid IMEI.

Instead:

1. Consult with firmware team on missing values in `asset_tracker_v2` for LwM2M,
2. if they cannot add them, update nRF Asset Tracker protocol (with
   a breaking change) to document that `asset_tracker_v2` using LwM2M does
   not send these values and make them optional.

This commit makes the unit tests fail, and this it is intentional. To make them
pass create a PR in the docs repo to get a schema that allows for the missing
values.