docs: link ADR 008 with a test
LwM2M has different versions. It is known that resources may or not be present on objects depending on the version.

For that reason, it is documented in ADR-008 which is the expected version of each object.

I feel the necessity to write a test about it because otherwise if versions of the '@nordicsemiconductor/lwm2m-types'
changes, there would be noticed only after something is broken and not immediately after the change.

What I would like to hear from you is if the 'non-functional-test' folder makes sense. For me, it is the place to set this test.
I don't see it in the converter test.