docs: link documentation with JSDocs
I think the documents (documents) created to see the data transition from
LwM2M to nRF Asset Tracker are useful to understand in a fast way how this converter works

I have a question about how this document should be referred to in the method.

I can use the link of Git Hub, but it is a doc that exists on the repo, so there is no need
to refer to it as an external resource. However, I am not pretty much convinced about the link
tag because I can not follow the link when clicking on it, so in that way, I think the GitHub link is better.

Which one feels better for you?