docs: metadata object was removed
The LwM2M objects used and their versions (see adr/008-default-lwm2m-version.md) support
timestamp as part of its interface (see adr/010-roam-timestamp-not-supported-by-lwm2m.md to check the only exception)

The reason why the metadata object was used is because of an issue with the reported
timestamp values in Azure, but this project (asset tracker LwM2M to nrf Asset Tracker)
is cloud agnostic. So makes no sense to add complexity if the scope of this project does not cover it.

The objects used here offer the option to report timestamp values and the issue of reported
 timestamp should be solved in the 'from Coiote Azure to LwM2M Asset Tracker' project.